Coordinate class work on

+++ SlicerInstruments.h
@@ -4,19 +4,19 @@
 class Instrument{
 	private:
 		std::string name;
-		float nozzle_radius_mm;
-		float feed_rate;		//xy-movement across material
-		float plunge_rate;	//z-movement across material
-		float allowable_surface_tilt_angle_rad;
+		double nozzle_radius_mm;
+		double feed_rate;		//xy-movement across material
+		double plunge_rate;	//z-movement across material
+		double allowable_surface_tilt_angle_rad;
 	public:
 		Instrument(){};
-		Instrument(std::string new_name, float new_radius, float new_feed_rate, float new_plunge_rate, float new_tilt){
+		Instrument(std::string new_name, double new_radius, double new_feed_rate, double new_plunge_rate, double new_tilt){
 			nozzle_radius_mm = new_radius;
 			feed_rate = new_feed_rate;
 			plunge_rate = new_plunge_rate;
 			allowable_surface_tilt_angle_rad = new_tilt;
 		}
-		float GetRadius() { return nozzle_radius_mm; }
+		double GetRadius() { return nozzle_radius_mm; }
 };
 
 class InstrumentsSet {
